Learning Commit
I’ve managed to add in the functionality of searching for an artist and
playing a clip from that artist to the index. This was ripped from the
“artistlookup.html” file that I’ve included. The purpose of doing this
was to demonstrate a GET after the user logs in. Next step will be to
replace the artist search with a list of the users playlists.