Bluemix and Design
Integrated Lala’s front end with the playlist loading that I’ve been
working on. If you want to run locally, its in the localhost-copy
folder. The portion of the repo that we push to bluemix is in the
second WoozyTunes folder. There’s a slight difference in the app.js
that makes the separate copies necessary