Playlist Buttons
Rather than automatically loading all the playlists, we now generate
buttons that when pressed load the playlist in a separate div. This is
important because this is what we will want to be doing when we connect
the buttons to the map portion of the interface.

Also: deleted some unnecessary old files and made small changes to the
css to get the “spotify style” buttons to work.